+++ site/src/app/sitemap.ts
@@ -1,5 +1,5 @@
-import type { MetadataRoute } from "next";
 import { config } from "@/config";
+import type { MetadataRoute } from "next";
 
 export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
   "use cache";