+++ src/data/__tests__/emoji.test.ts
@@ -8,35 +8,54 @@ describe("getEmojiData", () => {
   });
 
   it("should return the emoji data", async () => {
-    const data = await getEmojiData("en");
+    const data = await getEmojiData({ locale: "en" });
 
     expect(data).toBeDefined();
   });
 
   it("should support aborting the request", async () => {
     const controller = new AbortController();
-    const promise = getEmojiData(
-      "en",
-      Number.POSITIVE_INFINITY,
-      controller.signal,
-    );
+    const promise = getEmojiData({ locale: "en", signal: controller.signal });
 
     controller.abort();
 
-    const data = await promise;
-
-    expect(data).toBeDefined();
+    await expect(promise).rejects.toThrow(DOMException);
   });
 
-  it("should support a specific emoji version", async () => {
-    const data = await getEmojiData("en", 5);
+  it("should support a specific Emoji version", async () => {
+    const fetchSpy = vi.spyOn(globalThis, "fetch");
+    const data = await getEmojiData({ locale: "en", emojiVersion: 5 });
 
     expect(data).toBeDefined();
     expect(data.emojis.every((emoji) => emoji.version <= 5)).toBe(true);
+
+    expect(fetchSpy.mock.calls[0]?.[0]).toEqual(
+      "https://cdn.jsdelivr.net/npm/emojibase-data@5/en/data.json",
+    );
+    expect(fetchSpy.mock.calls[1]?.[0]).toEqual(
+      "https://cdn.jsdelivr.net/npm/emojibase-data@5/en/messages.json",
+    );
+  });
+
+  it("should support a custom Emojibase URL", async () => {
+    const fetchSpy = vi.spyOn(globalThis, "fetch");
+    const data = await getEmojiData({
+      locale: "en",
+      emojibaseUrl: "https://example.com/self-hosted-emojibase-data",
+    });
+
+    expect(data).toBeDefined();
+
+    expect(fetchSpy.mock.calls[0]?.[0]).toEqual(
+      "https://example.com/self-hosted-emojibase-data/en/data.json",
+    );
+    expect(fetchSpy.mock.calls[1]?.[0]).toEqual(
+      "https://example.com/self-hosted-emojibase-data/en/messages.json",
+    );
   });
 
   it("should save data locally", async () => {
-    await getEmojiData("en");
+    await getEmojiData({ locale: "en" });
 
     const localStorageData = localStorage.getItem(LOCAL_DATA_KEY("en"));
     const sessionStorageData = sessionStorage.getItem(SESSION_METADATA_KEY);
@@ -46,13 +65,13 @@ describe("getEmojiData", () => {
   });
 
   it("should use local data if available from a previous session", async () => {
-    await getEmojiData("en");
+    await getEmojiData({ locale: "en" });
 
     sessionStorage.clear();
 
     const fetchSpy = vi.spyOn(globalThis, "fetch");
 
-    await getEmojiData("en");
+    await getEmojiData({ locale: "en" });
 
     expect(fetchSpy).toHaveBeenCalledTimes(2);
     expect(fetchSpy.mock.calls[0]).toEqual([
@@ -69,7 +88,7 @@ describe("getEmojiData", () => {
     localStorage.setItem(LOCAL_DATA_KEY("en"), "{}");
     sessionStorage.setItem(SESSION_METADATA_KEY, "{}");
 
-    await getEmojiData("en");
+    await getEmojiData({ locale: "en" });
 
     const localStorageData = localStorage.getItem(LOCAL_DATA_KEY("en"));
     const sessionStorageData = sessionStorage.getItem(SESSION_METADATA_KEY);